move flash_attn and update dispatch_keys for flash_scaled_dot_prod
[ghstack-poisoned]
