+++ aten/src/ATen/native/native_functions.yaml
@@ -12482,7 +12482,7 @@
 - func: _flash_scaled_dot_product_self_attention(Tensor qkv, Tensor cumulative_sequence_length, int max_seqlen_batch, float dropout_p, bool causal) -> Tensor
   variants: function
   dispatch:
-    CPU, CUDA, NestedTensorCPU, NestedTensorCUDA: flash_scaled_dot_product_self_attention
+    CUDA, NestedTensorCUDA: flash_scaled_dot_product_self_attention
 
 - func: _transformer_decoder_only_layer_fwd(Tensor src, int embed_dim, int num_heads, Tensor qkv_weight, Tensor qkv_bias, Tensor proj_weight, Tensor proj_bias, bool use_gelu, bool norm_first, float eps, Tensor norm_weight_1, Tensor norm_bias_1, Tensor norm_weight_2, Tensor norm_bias_2, Tensor ffn_weight_1, Tensor ffn_bias_1, Tensor ffn_weight_2, Tensor ffn_bias_2, Tensor? mask=None, Tensor? incr_key=None, Tensor? incr_value=None) -> (Tensor, Tensor, Tensor)
   variants: function
