+++ aten/src/ATen/native/nested/cuda/flash_attn/gmem_tile.h
@@ -27,8 +27,8 @@
 
 #pragma once
 
-#include <ATen/native/nested/flash_attn/utils.h>
-#include <ATen/native/nested/flash_attn/gemm.h>
+#include <ATen/native/nested/cuda/flash_attn/utils.h>
+#include <ATen/native/nested/cuda/flash_attn/gemm.h>
 namespace fmha {
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
