+++ aten/src/ATen/native/nested/cuda/flash_attn/smem_tile.h
@@ -29,9 +29,9 @@
 
 
 
-#include <ATen/native/nested/flash_attn/gmem_tile.h>
-#include <ATen/native/nested/flash_attn/utils.h>
-#include <ATen/native/nested/flash_attn/gemm.h>
+#include <ATen/native/nested/cuda/flash_attn/gmem_tile.h>
+#include <ATen/native/nested/cuda/flash_attn/utils.h>
+#include <ATen/native/nested/cuda/flash_attn/gemm.h>
 
 namespace fmha {
